remove bottleneck in RunBOZORTH3
Originally the tables were normalized as in a relational databases.
This ended up creating a problem as the selection of probe and gallery
images used a cartesian product between the table of images and the
probe and gallery image IDs, causing a massive slowdown.

This change denormalizes the database, nesting each required object into
the data passed to the upstream function. Hence, the Mindtct object,
which depends on Image, stores not the foreign key but the entire Image
object itself, saving a lookup (this pattern is applied to Group and
Bozorth3 objects too). Space is cheaper than time in this case.

Small tests show a runtime reductino from several hours to a couple
minutes.